Move PCPCheckAndHeal to dotty.tools.dotc.staging
