+++ compiler/src/dotty/tools/dotc/core/StagingContext.scala
@@ -4,7 +4,7 @@ import dotty.tools.dotc.ast.tpd
 import dotty.tools.dotc.core.Contexts._
 import dotty.tools.dotc.ast.tpd
 import dotty.tools.dotc.util.Property
-import dotty.tools.dotc.transform.PCPCheckAndHeal
+import dotty.tools.dotc.staging.PCPCheckAndHeal
 
 object StagingContext {
