+++ compiler/src/dotty/tools/dotc/staging/PCPCheckAndHeal.scala
@@ -1,5 +1,5 @@
 package dotty.tools.dotc
-package transform
+package staging
 
 import dotty.tools.dotc.ast.{tpd, untpd}
 import dotty.tools.dotc.core.Annotations.BodyAnnotation
@@ -14,6 +14,7 @@ import dotty.tools.dotc.core.Types._
 import dotty.tools.dotc.util.SrcPos
 import dotty.tools.dotc.util.Spans._
 import dotty.tools.dotc.transform.SymUtils._
+import dotty.tools.dotc.transform.TreeMapWithStages
 import dotty.tools.dotc.typer.Checking
 import dotty.tools.dotc.typer.Implicits.SearchFailureType
 import dotty.tools.dotc.core.Annotations._
