+++ compiler/src/dotty/tools/dotc/transform/Splicing.scala
@@ -22,6 +22,7 @@ import dotty.tools.dotc.core.Names._
 import dotty.tools.dotc.core.StdNames._
 import dotty.tools.dotc.quoted._
 import dotty.tools.dotc.config.ScalaRelease.*
+import dotty.tools.dotc.staging.PCPCheckAndHeal
 
 import scala.annotation.constructorOnly
