+++ compiler/src/dotty/tools/dotc/transform/Staging.scala
@@ -12,8 +12,7 @@ import dotty.tools.dotc.core.Types._
 import dotty.tools.dotc.util.SrcPos
 import dotty.tools.dotc.transform.SymUtils._
 import dotty.tools.dotc.staging.StagingLevel.*
-
-
+import dotty.tools.dotc.staging.PCPCheckAndHeal
 
 /** Checks that the Phase Consistency Principle (PCP) holds and heals types.
  *
